+++ ext_localconf.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
+defined('TYPO3') or die();
+
 use Causal\Oidc\Hooks\DataHandlerOidc;
 use Causal\Oidc\LoginProvider\OidcLoginProvider;
 use Causal\Oidc\OidcConfiguration;
 use Causal\Oidc\Service\AuthenticationService;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-
-defined('TYPO3') or die();
 
 $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass'][] = DataHandlerOidc::class;
 
@@ -19,12 +19,12 @@ $settings = GeneralUtility::makeInstance(OidcConfiguration::class);
 
 // Service configuration
 $subTypes = array_merge(
-    ($settings->enableFrontendAuthentication) ? [
+    ($settings->hasProviderForFrontendAuthentication()) ? [
         'getUserFE',
         'authUserFE',
         'getGroupsFE',
     ] : [],
-    ($settings->enableBackendAuthentication) ? [
+    ($settings->hasProviderForBackendAuthentication()) ? [
         'getUserBE',
         'authUserBE',
     ] : [],
@@ -54,7 +54,7 @@ if (is_file($pharFileName)) {
     @include 'phar://' . $pharFileName . '/vendor/autoload.php';
 }
 
-if ($settings->enableBackendAuthentication) {
+if ($settings->hasProviderForBackendAuthentication()) {
     $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['backend']['loginProviders'][OidcLoginProvider::IDENTIFIER] = [
         'provider' => OidcLoginProvider::class,
         'sorting' => 50,