+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Feature: Support backend login.
 - Feature: Added Site Set
 - Method `getFreshAccessToken()` now actually returns the fresh access token.
+- Breaking: Extension Configuration is now in `/config/system/oidc.yml`. Example is in README.
+- Breaking: OIDC claim to TYPO3 user mapping is now configured in `oidc.yaml` as well. The former TypoScript configuration has been dropped.
 
 ## Version 4.0.0